Fix running configuration prior to installing.
This was specially problematic on git installs when no configuration directory argument was passed. We needed to trigger our default configuration function which required salt to be already checked out.

From now on, when a install_*_git_deps function is added, salt's git repository should be checked out at that stage.