FreeBSD update salt package -> sysutils/py-salt.
The salt package was renamed to the py- prefix to match other python
based ports. When built the name will be given a shortname of py27-salt,
in order to avoid Python versioning issues in the future we use the full
portname "sysutils/py-salt" which should always find the appropriate
version.