Alter bootstrap to handle new FreeBSD pkg changes
FreeBSD's `pkg` system is under development and has changed
it seems since this bootstrap was written.  A new install
of 9.2-Release will not work with the current version.

There are two notable changes

- use the "official" pkg.freebsd.org repository, which has support for more
  architectures (ll 2337).  the pcbsd site only supports amd64, and I at least
  still have some i386's :-)

- use the new format for repository conf files (ll 2349-2400) The format of conf
  files has changed and now needs a more YAML/JSON like definition, and the
  pkg.conf is deprecated and not needed.  I have replaced the SALTREPO as it
  seems out of date (Oct 2013)

This is tested on a clean install of 9.2-RELEASE and as the
very first file downloaded it then runs cleanly and
installs salt.

I hope contribution this helps your project