PR: Issue 394
When using Vagrant and Docker the Salt provisioner would fail as
Upstart is not running (and will most likely never be).

The easiest way to fix this is to get Docker to drop a file onto the
the filesystem (/tmp/disable_salt_checks) and then disable the service
checks from that.

Then to make it clear and optional from the command line I thought of
adding the -d argument option to do the same thing. If either of these
are present a warning message appears in the stdout.

For a Dockerfile you would add:
RUN touch /tmp/disable_salt_checks

To run from the command line you would do:
$ ./bootstrap-salt.sh -d

This adds the following var (defaults to $BS_FALSE):
_DISABLE_SALT_CHECKS