\| grouping doesn't work in POSIX sed
As part of #550, the reason FreeBSD falls on tags at all (fixed in #564) is because the REPO_URL sed substition line doesn't work in non-Linux sed. 

This fix replaces it with an equivalent POSIX compliant filter so it works on FreeBSD (and probably Solaris, Nexenta, etc.).

This was the only place in the script I could find where this occurs.