Make sure $STABLE_REV is set before setting up SaltStack repo
Fixes #882

When installing salt on Amazon Linux with a git reference, the $STABLE_REV
variable might not be set yet, causing malformed baseurl and gpgkey links
referenced in the /etc/yum.repos.d/saltstack-repo.repo file. We need to
ensure this variable is set before setting up the repo file.