SUSE: Refactor systemsmanagement_saltstack repo checks to be DRY
Also, relax the search for systemsmanagement_saltstack before
attempting to add and refresh the zypper repo list. When running
the bootstrap script on a Suse machine that already had the repo
present, we were failing the strict check, and then attempting
to add the repo even though it already existed. This would cause
the bootstrap script to exit and you couldn't try to install Salt
at a different version or try to run the script again.

Fixes #962