Run "pip install -U futures" for SmartOS git install
Fixes #764

When installing salt via git on SmartOS, salt commands would fail
because the required version of python futures wasn't installed.

We need to add a "pip install -U futures" call (triggered by passing
in -P) to the SmartOS git installation like we do for the required
version of tornado.