Update supported versions for Fedora: 23, 24, and 25.
The bootstrap script should match the supported versions of SaltStack
more closely, and we don't support versions that are considered EOL by
the OS distributors themselves.

This change updates the bootstrap script to support Fedora 23, 24, and 25.

Since "dnf" replaced "yum" as the package provider in Fedora 23, we no
longer need to set the $FEDORA_PACKAGE_MANAGER variable based on the
version of Fedora. This can simply be "dnf" now for Fedora installs.