Add "yakkety" to ubuntu_codename_translation function
Follow up to #998 and #976

This change will make it so the xenial repo.saltstack.com is not
added in `/etc/apt/sources.list.d/saltstack.list` file, but is
instead added as yakkety. The 16.10 link to repo.saltstack.com
currently will throw 404 errors, but the install will succeed
by installing 2016.3.1 from Ubuntu's repo. Git installs will work.

It's possible that a 16.10 repo will be added in the future. When
that happens, the bootstrap script will be ready.