Update bootstrap-salt.sh
This is a typo on line 4302 (applies to Alpine Linux) - missing leading / on the redirect to /dev/null.

Also, around line 4301, the rc.d scripts are never copied to /etc/init.d for rc-update to act on them. I've not located any OpenRC init scripts within the salt repo, so haven't fixed that.