Fix installing salt-cloud package on Arch Linux
When passing the ``-L`` option to the bootstrap script to install
the salt-cloud package, the ``apache-libcloud`` package must be
installed.

However, on Arch Linux this packages is named ``python2-apache-libcloud``.

This commit updates the name of the libcloud package to be installed
on Arch Linux so that the ``-L`` option can be executed successfully.

Fixes #1090