Wait for zypper processes to finish before calling zypper again
Fixes an issue where the bootstrap script bails out if there is a zypper
process running already on the machine to be bootstrapped. This can happen
when a machine has been modified to use zypper in a boot command. We need
to wait for the process to finish before calling the next zypper command.

Otherwise, the following error occurs:
```
System management is locked by the application with pid <pid#> (zypper).
```