FIX #1237 SmartOS should use pkgin show-deps
The install_smartos_deps function is used for both stable and git.
Only the m2crypt package is not included in the automatic dependacies
for the salt package.

We simply query the salt package to figure out the dependancies. That we
we have the right set for both git and stable. Even if pacakge names
change in the future.