cannot use 'if not set' expression here
_SALT_ETC_DIR=${BS_SALT_ETC_DIR:-/usr/local/etc/salt}
Says if variable is not set, set it to that.  By the time the script gets this far down, the variable is already set, so have to hardcode it to be /usr/local/etc/salt - part of the reason I had an 'if' statement around line 242, before the variable is set to default.