Pin previous release of ShellCheck in CI tests
The latest and greatest version of ShellCheck seems to have more
hardened rules. This currently breaks CI tests in lint phase and
shows error messages for untouched lines in a PR.

It would be better to freeze the SC version and update it explicitly
with fixing all new founded issues.