fix(opensuse): fix & add CI for openSUSE Tumbleweed
When `systemd` is installed on Tumbleweed, `systemctl` is actually
located at `/usr/bin/systemctl`.  It appears that the same applies for
openSUSE Leap except that a symlink is also provided there, i.e.
`/bin/systemctl` => `/usr/bin/systemctl`.

Tumbleweed installs Python 3.8, which was first supported in `3001` (Sodium).
Hence, testing on `2019.2` and `3000` is not possible.

When provisioning in Kitchen, another package is supplied compared to
openSUSE Leap:

* `systemd`: It appears that this is implicitly installed in Leap as a
  dependency of the `openssh` package; this doesn't apply in Tumbleweed,
  so it is installed explicitly.